requirements.txt
numpy>=1.24.0
scipy>=1.9.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.1
requests>=2.28.0
pyyaml>=6.0
tqdm>=4.65.0
holidays>=0.25
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
shapely>=2.0.0
geopandas>=0.12.0
folium>=0.14.0
osmnx>=1.2.0
plotly>=5.14.0
statsmodels>=0.14.0
pingouin>=0.5.3
sqlalchemy>=2.0.0
pymongo>=4.3.0
selenium>=4.9.0
jupyter>=1.0.0
ipykernel>=6.23.0
tensorboard>=2.13.0
wandb>=0.15.0
pytest>=7.3.0
pytest-cov>=4.1.0
black>=23.3.0
flake8>=6.0.0
fastapi>=0.95.0
uvicorn>=0.22.0
streamlit>=1.23.0
reportlab>=4.0.0
causalnex>=0.12.0
causal-learn>=0.1.3.3
dowhy>=0.9.0
econml>=0.14.0
